Rework custom migration to return migration SQL object, fix segfault
